Moved MainViewModel and added TestConfiguration class
