+++ shared/src/commonMain/kotlin/com/corrado4eyes/cucumberplayground/models/TestConfiguration.kt
@@ -1,0 +1,7 @@
+package com.corrado4eyes.cucumberplayground.models
+
+typealias TestConfigurationMap = Map<String, String>
+data class TestConfiguration(val configuration: TestConfigurationMap) {
+    val isLoggedIn = configuration["isLoggedIn"] == "true"
+    val testEmail = configuration["testEmail"] ?: "test@test.com"
+}
